chore: ensure page.getBySelector is available across tests
