+++ packages/components/src/tests/browser/setup.ts
@@ -1,5 +1,23 @@
 import { beforeAll, beforeEach, afterEach, vi } from "vitest";
 import { css } from "../../../support/formatting";
+import { type Locator, locators } from "vitest/browser";
+
+locators.extend({
+  getBySelector(selector: string): string {
+    return selector;
+  },
+});
+
+declare module "vitest/browser" {
+  interface LocatorSelectors {
+    /**
+     * Selects an element using a standard CSS selector.
+     *
+     * Note: prefer using more specific locators when possible for better test reliability.
+     */
+    getBySelector: (selector: string) => Locator;
+  }
+}
 
 beforeAll(() => {
   const style = document.createElement("style");
