+++ packages/components/src/tests/commonTests/browser/utils.ts
@@ -1,23 +1,5 @@
 import { camelCase } from "change-case";
 import { PascalCase } from "type-fest";
-import { type Locator, locators } from "vitest/browser";
-
-locators.extend({
-  getBySelector(selector: string): string {
-    return selector;
-  },
-});
-
-declare module "vitest/browser" {
-  interface LocatorSelectors {
-    /**
-     * Selects an element using a standard CSS selector.
-     *
-     * Note: prefer using more specific locators when possible for better test reliability.
-     */
-    getBySelector: (selector: string) => Locator;
-  }
-}
 
 type CalciteEventNamePrefix = `calcite${PascalCase<string>}`;
